Add button to upload model for extra networks
In the extra network section, adding a button for each tab
so that user can upload different types of models to different
directories.

TODO: The current implementation does not handle the case where
there are sub directories within each model folder.
