+++ html/extra-networks-upload-button.html
@@ -1,0 +1,5 @@
+<div class='card' style='white-space: nowrap; text-align: center; background-image: none; background-color:rgb(171 176 177 / 40%); {style}' onclick='{card_clicked}'>
+  <span class="helper" style="display: inline-block; height: 100%; vertical-align: middle;"></span>
+  <img id='{plus_sign_elem_id}' style='margin: auto; vertical-align: middle; display:inline-block' src='https://icons.veryicon.com/png/o/miscellaneous/standard-general-linear-icon/plus-60.png'>
+  <img id='{loading_sign_elem_id}' style='margin: auto; vertical-align: middle; display: none' src='https://i.gifer.com/origin/34/34338d26023e5515f6cc8969aa027bca_w200.gif'>
+</div>
