+++ html/footer.html
@@ -17,7 +17,7 @@
 
     <!-- Modal content -->
     <div id="notification-modal-content" class="modal-content">
-        <p style="color: #19974b">Notification.</p>
+        <p style="color: #0fd35d">Notification.</p>
     </div>
 
 </div>
@@ -36,7 +36,7 @@
 
 /* Modal Content */
 #notification-modal-content {{
-    background-color: rgba(68, 222, 148, 0.3);;
+    background-color: rgba(68, 222, 148, 0.5);
     margin: 0px auto;
     padding: 8px;
     text-align: center;
